# Refactor to improve code size, fix GCC 12