### Overview

Kinda a dump of (mostly) related changes I've been sitting on and never bothered to commit.

This is a set of changes to:
 - Reduce code size on `-O3`
    -  Especially on 32-bit, where `-O3` could get well over 100 kB `.text`
    - I feel this could still be improved if `hashLong` constant propagation can be reduced
 - Improve `-Os`/`XXH_SIZE_OPT` performance
 - Make things more clean and modular

**These haven't been micro-benchmarked yet.** The "what is worth inlining" is a rough estimate, and I want to go through the various recent compilers to find what works best. 


### Major changes:
 - GCC 12 will now be forced to `-O3` to work around the `-Og` bug. Fixes #800, #720 
    - Only applies when `__OPTIMIZE__`, `!__NO_INLINE__` and `XXH_SIZE_OPT <= 0` (`!__OPTIMIZE_SIZE__`), so `-O2` and `-Og` both apply
    - I tried with adjusting the inline hints but that tanked performance.
 - `XXH_INLINE`: Acts like existing `XXH_FORCE_INLINE` where it is disabled when `XXH_SIZE_OPT` is used.
    - `XXH_FORCE_INLINE` now actually force inlines unconditionally (when inline hints are on) and is used on XXH3 and XXH32's core loops and utilities. 
    - Greatly improves GCC `-Os` performance with only a small code size overhead
 - `XXH_32BIT_OPT`, `XXH_INLINE_64BIT`: These are used to reduce code bloat on 32-bit
    - This with the other related optimizations reduce `clang armv4t -O3` code size by **about half**, now being <100 kB
 - `XXH32` and `XXH64` have been rewritten to reuse the algorithm steps
 - `XXH3_update` has been rewritten to be much easier to understand and much smaller.
 - `XXH3_128bits_update` now tail calls `XXH3_64bits_update`, *greatly* reducing code size at the cost of one branch

### Minor changes:
 - GCC 11 and higher no longer force `-O2` on AVX2, that bug has been fixed.
 - `XXH32_state_t` and `XXH64_state_t` have some fields renamed and now use `unsigned char [N]` for their buffer
 - `XXH_OLD_NAMES` has been removed, it's been long enough
 - Fixed a few inline hints
 - Some reused snippets in XXH3 have been extracted into functions.
 - Fixed a theoretical integer overflow in the update code