@easyaspi314 2023-02-23
Seems like some apt repos are down which is causing some things to fail.