@easyaspi314 2023-02-23
How do the rest of these changes look aside from that? The changes shouldn't affect 64-bit codegen that much.

The biggest change codegen-wise is 32-bit not being forced to inline and unroll ugly functions as much and `-Os` being faster.