@Cyan4973 2023-02-26
I'm trying to get a sense of the scope of the changes in this PR.

To rephrase it :
- It's mostly focused on reducing the binary size, though speed might also be affected as a side effect
- It mostly concerns `-Os` compilation mode, performance modes `-O2` and `-O3` should not be negatively impacted
- It mostly concerns 32-bit compilation mode, 64-bit should not be negatively impacted
- Speed change is expected to be roughly neutral, except for 32-bit `-Os` where it's expected to be positive.
- Not sure if `arm` target is differently impacted than `x86`/`x64` one.

If that's the correct scope, then there are a number of measurements to do and publish to document this PR.

Longer term, I wonder if there is something that could / should be done to ensure these properties are preserved in the future. Otherwise, it would be easy for a future PR to inflate the binary size without anyone noticing it.