@Cyan4973 2023-02-27
Since the focus of this PR is on binary size, I wanted to have a more in-depth look at these evolutions.
In the following tables, I'm separating the variants `XXH32`, `XXH64` and `XXH128`, by producing limited library binaries.
For `XXH32`, the produced `libxxhash` only contains `XXH32` and some common functions.
For `XXH64`, the produced `libxxhash` has both previous content and `XXH64`, but no `XXH3` variant.
Finally, for `XXH128`, it's the full library.
The hope is that it will make it possible to observe the binary size changes of `XXH32` and `XXH64` separately, without mixing the (probably larger) changes due to vector extensions exploited by `XXH3` variants.
In the same vein, it's going to be more natural to add more vector extensions for comparisons.

Let's start this exercise with `XXH32` :

|  algo | target | compiler | flags | `dev` size | `pr802` size | lib size diff | 
| --- | --- | --- | --- | --- | --- | --- | 
| XXH32 | `x64` | `gcc` v9.4.0 | `-O3` | 16576 | 16556 | -20 | 
| XXH32 | `x64` | `gcc` v9.4.0 | `-O2` | 16616 | 16616 | 0 |
| XXH32 | `x64` | `gcc` v9.4.0 | `-Os` | 16600 | 16664 | +64 |
| XXH32 | `x86` | `gcc` v9.4.0 | `-m32 -O3` | 15456 | 15456 | 0 |
| XXH32 | `x86` | `gcc` v9.4.0 | `-m32 -O2` | 15488 | 15488 | 0 |
| XXH32 | `x86` | `gcc` v9.4.0 | `-m32 -Os` | 15440 | 15532 | +92 |

A few interesting details here.
To begin with, we know that performance remains largely unaffected for `XXH32`, staying at `6.9 GB/s` on this target cpu, whatever the compilation mode and flags.
We also note that the library size differences between modes and commits remain extremely small, which is consistent. 

Two remarquable trends are worth mentioning though : 
- Modes expected to produce smaller binary sizes actually produce _larger_ binary sizes. This trend is clearer for `pr802`. At the end of the story, the differences are not large, but that still goes _against_ expectations.
- `pr802` is especially worse at binary size for the `-Os` mode. This is also the reverse of my initial expectations.

It could be that binary size differences are in fact unrelated to `XXH32` proper, and due to minor differences in other generic symbols still compiled in the library, such as `XXH_versionNumber()` for example.

Now, onto `XXH64` :

|  algo | target | compiler | flags | `dev` size | `pr802` size | lib size diff | perf evol |
| --- | --- | --- | --- | --- | --- | --- | --- |
| XXH64 | `x64` | `gcc` v9.4.0 | `-O3` | 17072 | 17072 | 0 | 13.7 GB/s |
| XXH64 | `x64` | `gcc` v9.4.0 | `-O2` | 17016 | 17016 | 0 | 13.7 GB/s 
| XXH64 | `x64` | `gcc` v9.4.0 | `-Os` | 17040 | 17144 | +104 | 13.7 GB/s
| XXH64 | `x86` | `gcc` v9.4.0 | `-m32 -O3` | 19840 | 19872 | +32 | 3.1 -> 2.9 GB/s
| XXH64 | `x86` | `gcc` v9.4.0 | `-m32 -O2` | 19900 | 15924 | **-3976** | 2.8 -> 2.7 GB/s
| XXH64 | `x86` | `gcc` v9.4.0 | `-m32 -Os` | 19852 | 15916 | **-3936** | 2.7 -> 2.6 GB/s

Some learnings.
In `x64` mode, the library size difference is very small compared to the `XXH32` only library. Less than ~0.5KB is added. This seems to show that `XXH64` is compact enough, at all compilation settings. As we already know, speed remains unaffected, at `13.7 GB/s`, roughly double `XXH32`.
But in `x86` 32-bit mode, there is a more substantial binary size difference, and `dev` branch requires > +4 KB to add `XXH64`. This suggests that the `XXH64` algorithm gets expanded significantly. Probably some loop unrolling.
Binary size is improved by `pr802`, showing that it's possible to add `XXH64` with a budget of roughly ~0.5KB,  similar to `x64` mode.

We also know that `XXH64` performance is much lower in 32-bit mode, due to reliance on 64-bit multiply operations.
And unfortunately, `pr802` makes this situation slightly worse.
This would probably remain an acceptable trade-off when it saves almost ~4 KB of binary size, but strangely, at `-O3`, there is no such binary size saving, yet the performance difference is still very much present (checked several times).

It's unclear to me why there is a (small) performance difference between `dev` and `pr802` in 32-bit `x86` mode.
It's also unclear to me why the binary savings seems ineffective at `-O3` specifically, while `-O2` seems to benefit from it.