@Cyan4973 2023-02-27
The previous analysis was using Dynamic Library to compare generates binary sizes. It was assumed that this wouldn't make much difference compared to Static Libraries, with maybe just a constant factor added. 
This is incorrect : Static and Dynamic Library sizes are loosely correlated, but can evolve somewhat differently. 
This makes it difficult to decide which one is worth monitoring.
It could be argued that Static Library is closer to the concept of "generated binary size", so that would justify prefering Static library sizes. But this disregards the issue that these fragments of code, once integrated into a real executable (library or program) have to respect additional rules, such as code and data alignment. As a consequence, the impact of a change will be different depending on the target, and depending on the linking process. Therefore, it could also be argued that the Dynamic Library size gives a better view of the final impact of a change.

This is relatively difficult to make sense of, and difficult to decide, so I left both analysis in the thread. We can continue from the Static or the Dynamic library size analysis, depending on preference.

--------------------------------------

Since the focus of this PR is on binary size, I wanted to have a more in-depth look at these evolutions.
In the following tables, I'm separating the variants `XXH32`, `XXH64` and `XXH128`, by producing limited Static library binaries.
For `XXH32`, the produced `libxxhash` only contains `XXH32` and some common functions.
For `XXH64`, the produced `libxxhash` has both previous content and `XXH64`, but no `XXH3` variant.
Finally, for `XXH128`, it's the full library.
The hope is that it will make it possible to observe the binary size changes of `XXH32` and `XXH64` separately, without mixing the (probably larger) changes due to vector extensions exploited by `XXH3` variants.
In the same vein, it's going to be more natural to add more vector extensions for comparisons.

#### Let's start this exercise with `XXH32` :

|  algo | target | compiler | flags | `dev` size | `pr802` size | lib size diff | 
| --- | --- | --- | --- | --- | --- | ---:| 
| XXH32 | `x64` | `gcc` v9.4.0 | `-O3` | 4268 | 4404 | +136 |
| XXH32 | `x64` | `gcc` v9.4.0 | `-O2` | 4108 | 4228 | +120 |
| XXH32 | `x64` | `gcc` v9.4.0 | `-Os` | 3780 | 3836 | +56
| XXH32 | `x86` | `gcc` v9.4.0 | `-m32 -O3` | 4412 | 4460 | +48 |
| XXH32 | `x86` | `gcc` v9.4.0 | `-m32 -O2` | 4200 | 4248 | +48 |
| XXH32 | `x86` | `gcc` v9.4.0 | `-m32 -Os` | 3594 | 3590 | -4 |
| XXH32 | `aarch64` M1 | Apple `clang` v14.0 | `-O3` | 3416 | 3432 | +16 |
| XXH32 | `aarch64` M1 | Apple `clang` v14.0 | `-O2` | 3400 | 3416 | +16 |
| XXH32 | `aarch64` M1 | Apple `clang` v14.0 | `-Os` | 3392 | 3432 | +40 |

We know that `XXH32` speed remains largely unaffected, staying at `6.9 GB/s` on the `i7-9700k` no-turbo target cpu, whatever the compilation mode and flags. The story is essentially the same on `M1`, with `XXH32` delivering consistently `7.0 GB/s`, and a small (unimportant) degradation at `pr802` with `-Os` setting.

The library size differences between modes and commits remain small, which is unsurprising.
But we can also notice that `pr802` impact is _generally_ negative for binary size.
Not by a large amount, therefore it's not a big deal. 

Since there is no corresponding speed benefit, `pr802` looks globally _slightly_ negative for the `XXH32`-only scenario.
But as mentioned, this is minor, and can probably be ignored if there is a much larger benefit for other scenarios.

#### Now, onto `XXH64` :

|  algo | target | compiler | flags | `dev` size | `pr802` size | lib size diff | perf evol |
| --- | --- | --- | --- | --- | --- | ---:| --- |
| XXH64 | `x64` | `gcc` v9.4.0 | `-O3` | 7526 | 7886 | +360 | 13.7 GB/s |
| XXH64 | `x64` | `gcc` v9.4.0 | `-O2` | 7126 | 7446 | +320 | 13.7 GB/s 
| XXH64 | `x64` | `gcc` v9.4.0 | `-Os` | 6326 | 6406 | +80 | 13.7 GB/s
| XXH64 | `x86` | `gcc` v9.4.0 | `-m32 -O3` | 10638 | 9866 | -772 | 3.1 -> 2.9 GB/s
| XXH64 | `x86` | `gcc` v9.4.0 | `-m32 -O2` | 9110 | 7814 | **-1296** | 2.8 -> 2.7 GB/s
| XXH64 | `x86` | `gcc` v9.4.0 | `-m32 -Os` | 8048 | 6472 | **-1576** | 2.7 -> 2.6 GB/s
| XXH64 | `aarch64` M1 | Apple `clang` v14.0 | `-O3` | 5896 | 5912 | +16 | 14.0 GB/s |
| XXH64 | `aarch64` M1 | Apple `clang` v14.0 | `-O2` | 5832 | 5848 | +16 | 14.0 GB/s |
| XXH64 | `aarch64` M1 | Apple `clang` v14.0 | `-Os` | 5848 | 6128 | +280 | 14.0 -> 13.7 GB/s |

Some learnings:
As we already know, in 64-bit modes (`x64`, `aarch64`), speed remains unaffected, at `13.7 GB/s` on `i7-9700k`. It is only slightly degraded at `-Os` with `pr802` on `M1 Pro`, still offering roughly double `XXH32` speed.
Unfortunately, `pr802` doesn't offer any binary size savings for this scenario. It is actually _slightly bigger_ than `dev`.

On `x86` 32-bit mode, there is a more substantial binary size difference, and `dev` branch requires 2x binary size budget to add `XXH64`. This suggests that the `XXH64` algorithm gets expanded significantly. Probably some loop unrolling.
Binary size is improved by `pr802`, showing that it's possible to add `XXH64` with a budget roughly similar to `x64` mode.

We also know that `XXH64` performance is much lower in 32-bit mode, due to reliance on 64-bit multiply operations.
And unfortunately, `pr802` makes this situation slightly worse.
This would probably remain an acceptable trade-off when it saves binary size, but strangely, at `-O3`, binary size savings are not great, yet the performance difference is very much present (checked multiple times).

It's unclear to me why there is a (small) performance difference between `dev` and `pr802` in 32-bit `x86` mode.
It's also unclear to me why binary size savings is less effective at `-O3` specifically, while `-O2` seems to benefit more from it.

In conclusion, when it comes to `XXH64`, this `pr802` seems slightly negative for performance. There is an interesting binary size saving on `x86`, especially at `-O2` and `-Os` settings, that could potentially compensate the small speed degradation. But for `64-bit` modes, there is unfortunately no positive outcome to report .

#### `XXH128` :

Let's redo binary size measurements, using `static` library size instead.

|  algo | target | compiler | flags | `dev` size | `pr802` size | lib size diff | perf evol |
| --- | --- | --- | --- | --- | --- | --- | --- |
| XXH128 | `x64` + `SSE2` | `gcc` v9.4.0 | `-O3` | 67232 | 53200 | -14032 | 20.1 GB/s |
| XXH128 | `x64` + `SSE2` | `gcc` v9.4.0 | `-O2` | 40880 | 36864 | -4016 | 14.0 GB/s 
| XXH128 | `x64` + `SSE2` | `gcc` v9.4.0 | `-Os` | 19752 | 20432 | **+680** | 12.6 -> 14.9 GB/s
| XXH128 | `x86` + `scalar` | `gcc` v9.4.0 | `-m32 -O3` | 107280 | 73174 | -34106 | 5.7 -> 5.6 GB/s
| XXH128 | `x86` + `scalar` | `gcc` v9.4.0 | `-m32 -O2` | 60796 | 39188 | -21608 | 2.8 -> 2.7 GB/s
| XXH128 | `x86` + `scalar` | `gcc` v9.4.0 | `-m32 -Os` | 25152 | 23278 | -1878 | 2.2 -> 2.4 GB/s
| XXH128 | `aarch64` M1 + `NEON` | Apple `clang` v14.0 | `-O3` | 50456 | 44408 | -6048 | 35.8 GB/s |
| XXH128 | `aarch64` M1 + `NEON` | Apple `clang` v14.0 | `-O2` | 48768 | 42248 | -6520 | 35.8 GB/s |
| XXH128 | `aarch64` M1 + `NEON` | Apple `clang` v14.0 | `-Os` | 21600 | 24336 | **+2736** | 10.9 GB/s |

To begin with, it's clear that library size increases _significantly_ with `XXH3` variants. Sure, there are 2 variants, divided into a number of range strategies, plus streaming capabilities, ok. Yet, I'm nonetheless surprised that it requires so much. The `-Os` compilation mode is the only one that lies in the "acceptable" range.

Here, `pr802` brings pretty big binary size savings, at `-O2` and especially `-O3`.
For `-Os` though, it's less clear, and can actually *increase* size.

`-O3` remains an important compilation flag for best performance of `XXH128` on `gcc`: speed drops _very_ significantly with `-O2` and `-Os`.
On `clang`, the picture is slightly different : speed is exactly the same at `-O3` and `-O2`. But at `-Os`, the performance drop is *very* steep.

The performance impact of `pr802` is all over the place, from neutral, to slightly negative, to positive.
The most interesting aspect is seeing `-Os` mode receiving a speed boost on `gcc` + intel cpu.
For `x64`, this speed boost is significant enough to beat `-O2`, which is really nice since binary is also smaller.
For `x86`, the speed boost is not as important, but still a hefty +10% compared to `dev`.
So there's probably something  worth investigating there.

Note : some results are sometimes surprising. Ensuring that `pr802` is based on top of `dev` is also important to ensure this comparison analysis doesn't produce wrong conclusions.

To be continued : add more targets, more compilers, etc.